PRA-72 Обертка для homepage и endpage
