+++ packages/client/src/components/EndGame/index.ts
@@ -1,0 +1,1 @@
+export { EndGame } from './EndGame';
