+++ packages/client/src/components/WelcomeBox/index.ts
@@ -1,0 +1,1 @@
+export { WelcomeBox } from './WelcomeBox';
